Revert "error: switch float to double"
This reverts commit fe4bff55647f1a9981a7117cd02e3306409d17c9.

+++ Szeminarium1/Program.cs
@@ -107,7 +107,7 @@ namespace Szeminarium1
             uint vao = Gl.GenVertexArray();
             Gl.BindVertexArray(vao);
 
-            float[] vertexArray = new double[] {
+            float[] vertexArray = new float[] {
                 -0.5f, -0.5f, 0.0f,
                 +0.5f, -0.5f, 0.0f,
                  0.0f, +0.5f, 0.0f,
